+++ animation_nodes/sockets/sequence.py
@@ -1,6 +1,6 @@
 import bpy
 from bpy.props import *
-from bpy.types import Sequence
+from bpy.types import Strip
 from .. events import propertyChanged
 from .. base_types import AnimationNodeSocket, PythonListSocket
 
@@ -54,7 +54,7 @@ class SequenceSocket(bpy.types.NodeSocket, AnimationNodeSocket):
 
     @classmethod
     def correctValue(cls, value):
-        if isinstance(value, Sequence) or value is None:
+        if isinstance(value, Strip) or value is None:
             return value, 0
         return cls.getDefaultValue(), 2
 
@@ -75,6 +75,6 @@ class SequenceListSocket(bpy.types.NodeSocket, PythonListSocket):
     @classmethod
     def correctValue(cls, value):
         if isinstance(value, list):
-            if all(isinstance(element, Sequence) or element is None for element in value):
+            if all(isinstance(element, Strip) or element is None for element in value):
                 return value, 0
         return cls.getDefaultValue(), 2